pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Gabx-dev <58875500+Gabx-dev@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.2"
Django = "^4.1"
djangorestframework = "^3.13.1"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
django-extensions = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
